[Improvement-3369][api] Introduce taskrecord, udffunc, workflowlineage and workergroup service interface for clear code
